增加图片渲染

+++ core/src/main/java/top/untoldstudio/rimeui/core/MathTool.java
@@ -4,7 +4,7 @@ public final class MathTool {
     public static int round(double value){
         return (int)Math.round(value);
     }
-    public static float round(float value){
-        return (float)Math.round(value);
+    public static int round(float value){
+        return Math.round(value);
     }
 }
